requirements.txt
aiohttp>=3.9.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
prometheus-client>=0.19.0
pyjwt>=2.8.0
cryptography>=41.0.7
